app/auth.py
# app/auth.py
from datetime import datetime, timedelta, timezone
from typing import Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from passlib.context import CryptContext
from sqlalchemy.orm import Session

from app.core.settings import get_settings
from app.database import get_db
from app.models import Pessoa

settings = get_settings()

# Hash/verify de senha
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
def hash_password(password: str) -> str:
    return pwd_context.hash(password)
def verify_password(plain: str, hashed: str) -> bool:
    return bool(hashed) and pwd_context.verify(plain, hashed)

# ⚠️ tem que bater com o seu endpoint real de login (/token)
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# JWT helpers
def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (
        expires_delta or timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)

def decode_access_token(token: str) -> dict:
    return jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])

# Dependência para pegar o usuário autenticado
async def get_current_pessoa(
    token: str = Depends(oauth2_scheme),
    db: Session = Depends(get_db),
) -> Pessoa:
    cred_exc = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Não autenticado",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = decode_access_token(token)
        email: Optional[str] = payload.get("sub")
        if not email:
            raise cred_exc
    except JWTError:
        raise cred_exc

    pessoa = db.query(Pessoa).filter(Pessoa.email == email).first()
    if not pessoa:
        raise cred_exc
    return pessoa
